Make system directory check only for CD games
System directory will never be NULL anyways since RETRO_ENVIRONMENT_GET_SYSTEM_DIRECTORY
will always return the system-directory preset or content directory.
This is only needed for CD games when it loads the bios file. So
instead of early fail if ever the system directory will be NULL, this
will now only fail during loading a CDROM game and when trying to load
the bios file.